app/requirements.txt
Flask
mysql-connector
simplejson
python-dotenv
requests
beautifulsoup4